Add CMP server demo and instructions
