+++ cmp/bouncycastle/build.gradle.kts
@@ -10,12 +10,6 @@ repositories {
 }
 
 dependencies {
-//    testImplementation(platform("org.junit:junit-bom:5.9.1"))
-//    testImplementation("org.junit.jupiter:junit-jupiter")
     implementation("org.bouncycastle:bcprov-jdk18on:1.73")
     implementation("org.bouncycastle:bcpkix-jdk18on:1.73")
-}
-
-tasks.test {
-    useJUnitPlatform()
 }
